move ForRange to ast
