+++ compiler/noirc_frontend/src/ast/statement.rs
@@ -1,9 +1,14 @@
 use std::fmt::Display;
+use std::sync::atomic::{AtomicU32, Ordering};
 
 use crate::lexer::token::SpannedToken;
-use crate::parser::{ForRange, ParserError, ParserErrorReason};
+use crate::parser::{ParserError, ParserErrorReason};
 use crate::token::Token;
-use crate::{Expression, ExpressionKind, IndexExpression, MemberAccessExpression, UnresolvedType};
+use crate::{
+    BlockExpression, Expression, ExpressionKind, IndexExpression, MemberAccessExpression,
+    MethodCallExpression, UnresolvedType,
+};
+use acvm::FieldElement;
 use iter_extended::vecmap;
 use noirc_errors::{Span, Spanned};
 
@@ -479,19 +484,122 @@ impl LValue {
 }
 
 #[derive(Debug, PartialEq, Eq, Clone)]
+pub enum ForRange {
+    Range(/*start:*/ Expression, /*end:*/ Expression),
+    Array(Expression),
+}
+
+impl ForRange {
+    /// Create a 'for' expression taking care of desugaring a 'for e in array' loop
+    /// into the following if needed:
+    ///
+    /// {
+    ///     let fresh1 = array;
+    ///     for fresh2 in 0 .. std::array::len(fresh1) {
+    ///         let elem = fresh1[fresh2];
+    ///         ...
+    ///     }
+    /// }
+    pub(crate) fn into_for(
+        self,
+        identifier: Ident,
+        block: Expression,
+        for_loop_span: Span,
+    ) -> StatementKind {
+        /// Counter used to generate unique names when desugaring
+        /// code in the parser requires the creation of fresh variables.
+        /// The parser is stateless so this is a static global instead.
+        static UNIQUE_NAME_COUNTER: AtomicU32 = AtomicU32::new(0);
+
+        match self {
+            ForRange::Range(..) => {
+                unreachable!()
+            }
+            ForRange::Array(array) => {
+                let array_span = array.span;
+                let start_range = ExpressionKind::integer(FieldElement::zero());
+                let start_range = Expression::new(start_range, array_span);
+
+                let next_unique_id = UNIQUE_NAME_COUNTER.fetch_add(1, Ordering::Relaxed);
+                let array_name = format!("$i{next_unique_id}");
+                let array_span = array.span;
+                let array_ident = Ident::new(array_name, array_span);
+
+                // let fresh1 = array;
+                let let_array = Statement {
+                    kind: StatementKind::Let(LetStatement {
+                        pattern: Pattern::Identifier(array_ident.clone()),
+                        r#type: UnresolvedType::unspecified(),
+                        expression: array,
+                    }),
+                    span: array_span,
+                };
+
+                // array.len()
+                let segments = vec![array_ident];
+                let array_ident =
+                    ExpressionKind::Variable(Path { segments, kind: PathKind::Plain });
+
+                let end_range = ExpressionKind::MethodCall(Box::new(MethodCallExpression {
+                    object: Expression::new(array_ident.clone(), array_span),
+                    method_name: Ident::new("len".to_string(), array_span),
+                    arguments: vec![],
+                }));
+                let end_range = Expression::new(end_range, array_span);
+
+                let next_unique_id = UNIQUE_NAME_COUNTER.fetch_add(1, Ordering::Relaxed);
+                let index_name = format!("$i{next_unique_id}");
+                let fresh_identifier = Ident::new(index_name.clone(), array_span);
+
+                // array[i]
+                let segments = vec![Ident::new(index_name, array_span)];
+                let index_ident =
+                    ExpressionKind::Variable(Path { segments, kind: PathKind::Plain });
+
+                let loop_element = ExpressionKind::Index(Box::new(IndexExpression {
+                    collection: Expression::new(array_ident, array_span),
+                    index: Expression::new(index_ident, array_span),
+                }));
+
+                // let elem = array[i];
+                let let_elem = Statement {
+                    kind: StatementKind::Let(LetStatement {
+                        pattern: Pattern::Identifier(identifier),
+                        r#type: UnresolvedType::unspecified(),
+                        expression: Expression::new(loop_element, array_span),
+                    }),
+                    span: array_span,
+                };
+
+                let block_span = block.span;
+                let new_block = BlockExpression(vec![
+                    let_elem,
+                    Statement { kind: StatementKind::Expression(block), span: block_span },
+                ]);
+                let new_block = Expression::new(ExpressionKind::Block(new_block), block_span);
+                let for_loop = Statement {
+                    kind: StatementKind::For(ForLoopStatement {
+                        identifier: fresh_identifier,
+                        range: ForRange::Range(start_range, end_range),
+                        block: new_block,
+                        span: for_loop_span,
+                    }),
+                    span: for_loop_span,
+                };
+
+                let block = ExpressionKind::Block(BlockExpression(vec![let_array, for_loop]));
+                StatementKind::Expression(Expression::new(block, for_loop_span))
+            }
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Clone)]
 pub struct ForLoopStatement {
     pub identifier: Ident,
     pub range: ForRange,
     pub block: Expression,
     pub span: Span,
-}
-
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct ForRangeLoopStatement {
-    pub identifier: Ident,
-    pub start_range: Expression,
-    pub end_range: Expression,
-    pub block: Expression,
 }
 
 impl Display for StatementKind {
