+++ compiler/noirc_frontend/src/parser/mod.rs
@@ -11,28 +11,19 @@ mod labels;
 #[allow(clippy::module_inception)]
 mod parser;
 
-use std::sync::atomic::{AtomicU32, Ordering};
-
 use crate::token::{Keyword, Token};
 use crate::{ast::ImportStatement, Expression, NoirStruct};
 use crate::{
-    BlockExpression, ExpressionKind, ForLoopStatement, Ident, IndexExpression, LetStatement,
-    MethodCallExpression, NoirFunction, NoirTrait, NoirTraitImpl, NoirTypeAlias, Path, PathKind,
-    Pattern, Recoverable, Statement, StatementKind, TypeImpl, UnresolvedType, UseTree,
+    Ident, LetStatement, NoirFunction, NoirTrait, NoirTraitImpl, NoirTypeAlias, Recoverable,
+    StatementKind, TypeImpl, UseTree,
 };
 
-use acvm::FieldElement;
 use chumsky::prelude::*;
 use chumsky::primitive::Container;
 pub use errors::ParserError;
 pub use errors::ParserErrorReason;
 use noirc_errors::Span;
 pub use parser::parse_program;
-
-/// Counter used to generate unique names when desugaring
-/// code in the parser requires the creation of fresh variables.
-/// The parser is stateless so this is a static global instead.
-static UNIQUE_NAME_COUNTER: AtomicU32 = AtomicU32::new(0);
 
 #[derive(Debug, Clone)]
 pub(crate) enum TopLevelStatement {
@@ -475,112 +466,6 @@ impl Precedence {
     /// are '+' and '-' with precedence Sum.
     fn lowest_type_precedence() -> Self {
         Precedence::Sum
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub enum ForRange {
-    Range(/*start:*/ Expression, /*end:*/ Expression),
-    Array(Expression),
-}
-
-impl ForRange {
-    /// Create a 'for' expression taking care of desugaring a 'for e in array' loop
-    /// into the following if needed:
-    ///
-    /// {
-    ///     let fresh1 = array;
-    ///     for fresh2 in 0 .. std::array::len(fresh1) {
-    ///         let elem = fresh1[fresh2];
-    ///         ...
-    ///     }
-    /// }
-    pub(crate) fn into_for(
-        self,
-        identifier: Ident,
-        block: Expression,
-        for_loop_span: Span,
-    ) -> StatementKind {
-        match self {
-            ForRange::Range(..) => {
-                unreachable!()
-            }
-            ForRange::Array(array) => {
-                let array_span = array.span;
-                let start_range = ExpressionKind::integer(FieldElement::zero());
-                let start_range = Expression::new(start_range, array_span);
-
-                let next_unique_id = UNIQUE_NAME_COUNTER.fetch_add(1, Ordering::Relaxed);
-                let array_name = format!("$i{next_unique_id}");
-                let array_span = array.span;
-                let array_ident = Ident::new(array_name, array_span);
-
-                // let fresh1 = array;
-                let let_array = Statement {
-                    kind: StatementKind::Let(LetStatement {
-                        pattern: Pattern::Identifier(array_ident.clone()),
-                        r#type: UnresolvedType::unspecified(),
-                        expression: array,
-                    }),
-                    span: array_span,
-                };
-
-                // array.len()
-                let segments = vec![array_ident];
-                let array_ident =
-                    ExpressionKind::Variable(Path { segments, kind: PathKind::Plain });
-
-                let end_range = ExpressionKind::MethodCall(Box::new(MethodCallExpression {
-                    object: Expression::new(array_ident.clone(), array_span),
-                    method_name: Ident::new("len".to_string(), array_span),
-                    arguments: vec![],
-                }));
-                let end_range = Expression::new(end_range, array_span);
-
-                let next_unique_id = UNIQUE_NAME_COUNTER.fetch_add(1, Ordering::Relaxed);
-                let index_name = format!("$i{next_unique_id}");
-                let fresh_identifier = Ident::new(index_name.clone(), array_span);
-
-                // array[i]
-                let segments = vec![Ident::new(index_name, array_span)];
-                let index_ident =
-                    ExpressionKind::Variable(Path { segments, kind: PathKind::Plain });
-
-                let loop_element = ExpressionKind::Index(Box::new(IndexExpression {
-                    collection: Expression::new(array_ident, array_span),
-                    index: Expression::new(index_ident, array_span),
-                }));
-
-                // let elem = array[i];
-                let let_elem = Statement {
-                    kind: StatementKind::Let(LetStatement {
-                        pattern: Pattern::Identifier(identifier),
-                        r#type: UnresolvedType::unspecified(),
-                        expression: Expression::new(loop_element, array_span),
-                    }),
-                    span: array_span,
-                };
-
-                let block_span = block.span;
-                let new_block = BlockExpression(vec![
-                    let_elem,
-                    Statement { kind: StatementKind::Expression(block), span: block_span },
-                ]);
-                let new_block = Expression::new(ExpressionKind::Block(new_block), block_span);
-                let for_loop = Statement {
-                    kind: StatementKind::For(ForLoopStatement {
-                        identifier: fresh_identifier,
-                        range: ForRange::Range(start_range, end_range),
-                        block: new_block,
-                        span: for_loop_span,
-                    }),
-                    span: for_loop_span,
-                };
-
-                let block = ExpressionKind::Block(BlockExpression(vec![let_array, for_loop]));
-                StatementKind::Expression(Expression::new(block, for_loop_span))
-            }
-        }
     }
 }
